+++ include/Ark/Ark.hpp
@@ -7,12 +7,4 @@
 #include <Ark/VM/VM.hpp>
 #include <Ark/Compiler/Compiler.hpp>
 
-namespace Ark
-{
-    void compile(bool debug, const std::string& file, const std::string& output);
-    void vm(bool debug, const std::string& file);
-    void bcr(const std::string& file);
-    void run(const std::string& file, bool debug, bool recompile=false);
-}
-
 #endif  // ark_ark